third_party/WebKit/LayoutTests/http/tests/devtools/service-workers/lazy-addeventlisteners-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that a warning is shown in the console if addEventListener is called after initial evaluation of the service worker script.

Message count: 1
service-worker-lazy-…deventlistener.js:2 Event handler of 'install' event must be added on the initial evaluation of worker script.
setTimeout @ service-worker-lazy-…deventlistener.js:2
setTimeout (async)
(anonymous) @ service-worker-lazy-…deventlistener.js:1
